OFA/requirements.txt
-e ./fairseq/
opencv-python
timm
ftfy==6.0.3
tensorboardX==2.4.1
pycocotools==2.0.4
pycocoevalcap==1.2
pytorch_lightning
einops
datasets
rouge_score
pandas
tqdm
pillow
wandb
fairseq
bert_score
fastapi
matplotlib
python-multipart
rich